refactor(cli): move extractor/nlsearch to client, wire callers to the interface
