+++ cli/cmd/import/config.go
@@ -13,7 +13,7 @@ import (
 	enricherconfig "github.com/agntcy/dir-importer/enricher/config"
 	"github.com/agntcy/dir-importer/enricher/toolhost"
 	"github.com/agntcy/dir-importer/transformer"
-	internalextractor "github.com/agntcy/dir/cli/internal/extractor"
+	"github.com/agntcy/dir/client/extractor"
 	sdk "github.com/agntcy/oasf-sdk/pkg/extractor"
 	"github.com/spf13/pflag"
 	"gopkg.in/yaml.v3"
@@ -287,7 +287,7 @@ func toToolHostConfig(fc toolHostFileConfig) toolhost.Config {
 // `dirctl init`; explicit values override the saved config.
 func loadExtractorFromConfig(fc *extractorEnricherFileConfig) (*sdk.Extractor, error) {
 	if fc.OASFUrl == "" && fc.AssetDir == "" {
-		ext, err := internalextractor.LoadConfigured()
+		ext, err := extractor.LoadConfigured()
 		if err != nil {
 			return nil, fmt.Errorf("load configured extractor: %w", err)
 		}
@@ -295,7 +295,7 @@ func loadExtractorFromConfig(fc *extractorEnricherFileConfig) (*sdk.Extractor, e
 		return ext, nil
 	}
 
-	ext, err := internalextractor.Load(internalextractor.Config{
+	ext, err := extractor.Load(extractor.Config{
 		OASFURL:  fc.OASFUrl,
 		AssetDir: fc.AssetDir,
 	})
