+++ client/extractor/assets_test.go
