+++ client/extractor/config_test.go
