+++ client/extractor/provision_integration_test.go
